eval: fix cast from map entry to slot index
Casting (void *) to (unsigned int) is triggering on 32 bit due to
differing sizes.

It is okay in this case, the value is never expected to be big enough
to have truncation at this point, but then, using (unsigned long) is
good as well and shuts up the warning.

+++ sheep/eval.c
@@ -49,9 +49,9 @@ static sheep_t hash(struct sheep_vm *vm, sheep_t container,
 		goto err;
 
 	if (value)
-		slots[(unsigned int)entry] = value;
+		slots[(unsigned long)entry] = value;
 
-	return slots[(unsigned int)entry];
+	return slots[(unsigned long)entry];
 err:
 	obj = sheep_repr(container);
 	fprintf(stderr, "can not find %s in %s\n", key, obj);
